Implement SerializeBinary for SerializableFst types
`SerializableFst` trait already needs its type to be converted to binary data. Therefore, adding `SerializeBinary` requirement for the trait makes it possible to add default implementations of `SerializableFst::read` and `SerializableFst::write`. In addition, having an API to convert an FST to/from binary data may be useful for other serialization implementations.